Only create on-disk LMDB navigator environment on-demand
Moves the creation of the LMDB environment to the `finalize()`
function where we write out the navigator data to disk.

This means we only write to disk when is actually needed and if
`finalize()` is never called, we won't unnecessairly create an
an empty LMDB database on disk.

This also prepares for a world where we have multiple output formats
of navigator index data.