Fix ResolvedTopicReference memory leak
The caching mechanism in ResolvedTopicReference never releases
references to the ResolvedTopicReferences it holds.

This has likely always been a problem since the introduction of the
cache, but is just now appearing as a memory leak when profile SwiftDocC
because we recently added copy-on-write functionality to
`ResolvedTopicReference`.

It’s important to release unused `ResolvedTopicReference`s because
We can create potentially tens of thousands of them when compiling large
documentation catalogs or loading large navigator indexes.

Resolves rdar://86035019.