SE-0356: Implement snippet slicing
Add an optional `slice` argument to the `@Snippet` directive, which
will only render a code listing comprised of the line range of the slice.

Remove snippet groups, as they are no longer used in the current design.

Replace logic that filters snippets and fake modules that hold snippets
to a more generic "is virtual" logic. A virtual symbol, topic, or documentation
node should not have its own page of documentation.

Testing: Moved snippet tests to their own test bundle as it was causing a lot
of thrash with navigator tests.

rdar://95220570

# Conflicts:
#	Sources/SwiftDocC/Infrastructure/DocumentationContext.swift
#	Sources/SwiftDocC/Semantics/MarkupReferenceResolver.swift
#	Tests/SwiftDocCTests/Infrastructure/DocumentationContext/DocumentationContextTests.swift
#	Tests/SwiftDocCTests/Test Bundles/TestBundle.docc/Test-snippets.symbols.json