Revert changes to 'DocumentationContext.swift' from "`@PageKind` sets the kind for the documentation node (#626)"
This reverts the changes to 'DocumentationContext.swift' from
commit 792b9e8f5b324640d8d930e4264c847fb32a6760.

This change introduced a regression where sample code pages
we're always automatically curated even if they'd been manually
curated. This made it impossible to curate a sample code
page without ending up with duplicate entries for that page.

Only the changes to 'DocumentationContext.swift' are being reverted
so that the tests added for the other bug this change fixed remain.

rdar://112223641